Address systemd gate review findings (#28)
Harden the systemd-gate checkout, run the gate through Cargo so custom
target directories work, and move duplicated daemon validation helpers
into `repovec-core`.
